refactor(ScorePolicy): 점수 계산 인터페이스 삭제
